Пишу на русском потому что дохуя тут)) Разбил все вьюхи и сериалайзеры по папкам протестил все урлы кроме урлов для приложения файлов, все работают теперь, манагерова тоже можно добавлять, короче удачи тебе разбирайся что тут у меня)))
