+++ src/trello/serializers/attachments_serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from trello.models import *
+
+
+class AttachmentDetailSerializer(serializers.ModelSerializer):
+	class Meta:
+		model = Attachment
+		fields = '__all__'
