+++ src/trello/serializers/subtask_serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from trello.models import *
+
+
+class SubtaskDetailSerializer(serializers.ModelSerializer):
+	class Meta:
+		model = Subtask
+		fields = '__all__'
