+++ src/trello/serializers/user_serilizer.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from trello.models import *
+
+
+class UserDetailSerializer(serializers.ModelSerializer):
+	class Meta:
+		model = User
+		fields = '__all__'
